+++ NEWS.md
@@ -41,6 +41,8 @@
 * `format_bf()` with `stars = TRUE` used the `°` symbol for inferiority
   (evidence *against* the comparison).
 
+* Added support for `coxph.panel` models.
+
 * Overhaul of documentation for the package-functions.
 
 ## Bug fix